% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datalist.R
\docType{data}
\name{datalist}
\alias{datalist}
\title{Data extracted from the initial metatranscriptomic data}
\format{
A data frame with 53940 rows and 10 variables:
\describe{
\item{table.data}{table with raw counts and the corresponding gene IDs}
\item{file.species}{a table with the gene IDs corresponding to the species}
\item{design}{a table of the experience's design, with all the parameters}
}
}
\usage{
datalist
}
\description{
A dataset containing a table with raw counts and the corresponding gene IDs,
a table with the gene IDs corresponding to the species
a table of the experience's design, with all the parameters
}
\keyword{datasets}
