% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecodif_estidispe.R
\name{ecodif_estidispe}
\alias{ecodif_estidispe}
\title{ecodif_estidispe function
calculate the dispersion factors}
\usage{
ecodif_estidispe(obj.dds)
}
\arguments{
\item{obj.dds}{obj.dds, list of tables, parameters and datas}
}
\value{
obj.dds, list of tables, parameters and datas
}
\description{
ecodif_estidispe function
calculate the dispersion factors
}
