% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecodif_wald.R
\name{ecodif_wald}
\alias{ecodif_wald}
\title{ecodif_wald function
calculate Wald statistics}
\usage{
ecodif_wald(obj.dds, ...)
}
\arguments{
\item{obj.dds}{obj.dds, list of tables, parameters and datas}

\item{...}{arguments to be passed to DESeq2::results}
}
\value{
obj.dds, list of tables, parameters and datas
}
\description{
ecodif_wald function
calculate Wald statistics
}
