% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_ecodif.R
\name{main_ecodif}
\alias{main_ecodif}
\title{Main function
Contains all the functions}
\usage{
main_ecodif(table.data, file.species, design)
}
\arguments{
\item{table.data}{raw counts matrix}

\item{file.species}{matrix with species}

\item{design}{design decided by the user}
}
\value{
obj.dds, list of tables, parameters and datas
}
\description{
Main function
Contains all the functions
}
